cov: changing size of nwatches to prevent truncation.

+++ clamd/onaccess_ddd.h
@@ -38,7 +38,7 @@ struct ddd_thrarg {
 };
 
 static int onas_ddd_init_ht(uint32_t ht_size);
-static int onas_ddd_init_wdlt(uint32_t nwatches);
+static int onas_ddd_init_wdlt(uint64_t nwatches);
 static int onas_ddd_grow_wdlt();
 
 static int onas_ddd_watch(const char *pathname, int fan_fd, uint64_t fan_mask, int in_fd, uint64_t in_mask);
@@ -46,7 +46,7 @@ static int onas_ddd_watch_hierarchy(const char* pathname, size_t len, int fd, ui
 static int onas_ddd_unwatch(const char *pathname, int fan_fd, int in_fd);
 static int onas_ddd_unwatch_hierarchy(const char* pathname, size_t len, int fd, uint32_t type);
 
-int onas_ddd_init(uint32_t nwatches, size_t ht_size);
+int onas_ddd_init(uint64_t nwatches, size_t ht_size);
 void *onas_ddd_th(void *arg);
 static void onas_ddd_exit(int sig);
